+++ CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix a crash when a subpipeline is passed as an operation argument
 
+- bash completion: fix mid-word completion behaviour, update docs
+
 ## [0.21.1] - 2026-06-14
 
 ### Fixed